profile: restore global hiding of all events
26f20b805d3c made the event hiding mechanism more fine grained
(by considering also the severity). The commit inadvertently
removed the feature that "hide similar events" would hide the
events from all dives.

Restore this. It's a bit complex, because the "eventname" now
has to remember the severity. Therefore, change "eventname"
into "eventtype".

The UI is a bit inconsistent, as "hide similar events" affects
all dives and is remembered across switching dives. In contrast,
"hide event" is not remembered. This needs some rework.

Signed-off-by: Berthold Stoeger <bstoeger@mail.tuwien.ac.at>